updated pyscripts
